requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==4.2.21
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.3.0
fonttools==4.58.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
kiwisolver==1.4.7
kombu==5.5.3
matplotlib==3.8.3
multidict==6.4.3
openai==0.28.1
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
zipp==3.21.0